+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - CMS-1666 - Upgraded directory-components to version 20.0.0 (Updated breadcrumbs tags and GA360 stuff to reflect the latest changes)
 - CMS-1666 - Upgraded directory-constants to 18.0.0
 - XOT-914 - Add automated tests for search sort order
+- XOT-906 - Update wording on search feedbac form
 
 ### Fixed bugs
 - TT-1602 - Fix radio buttons on /contact/domestic/